user_service/requirements.txt
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
gunicorn==21.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
urllib3
Werkzeug==3.1.3
pytest==8.3.4
Flask-SocketIO
eventlet
flask_session
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-requests
opentelemetry-propagator-b3
opentelemetry-propagator-jaeger